phase-1/jetson/requirements.txt
pyyaml==6.0.2
pyserial==3.5
